// interfaces/feed.tsx
export default interface FeedInterface {
  copyright: string;
  gamePk:    number;
  link:      string;
  metaData:  MetaData;
  gameData:  GameData;
  liveData:  LiveData;
}

export interface GameData {
  game:              Game;
  datetime:          Datetime;
  status:            Status;
  teams:             GameDataTeams;
  players:           { [key: string]: Player };
  venue:             Venue;
  officialVenue:     OfficialVenue;
  weather:           Weather;
  gameInfo:          GameInfo;
  review:            Review;
  flags:             Flags;
  alerts:            any[];
  probablePitchers:  ProbablePitchers;
  officialScorer:    OfficialScorer;
  primaryDatacaster: OfficialScorer;
}

export interface Datetime {
  dateTime:     Date;
  originalDate: Date;
  officialDate: Date;
  dayNight:     string;
  time:         string;
  ampm:         string;
}

export interface Flags {
  noHitter:            boolean;
  perfectGame:         boolean;
  awayTeamNoHitter:    boolean;
  awayTeamPerfectGame: boolean;
  homeTeamNoHitter:    boolean;
  homeTeamPerfectGame: boolean;
}

export interface Game {
  pk:              number;
  type:            AbstractGameCodeEnum;
  doubleHeader:    DoubleHeader;
  id:              string;
  gamedayType:     GamedayType;
  tiebreaker:      DoubleHeader;
  gameNumber:      number;
  calendarEventID: string;
  season:          string;
  seasonDisplay:   string;
}
// ...
export interface GameInfo {
  attendance:          number;
  firstPitch:          Date;
  gameDurationMinutes: number;
}

export interface OfficialScorer {
  id:       number;
  fullName: string;
  link:     string;
}

export interface OfficialVenue {
  id:   number;
  link: string;
}

export interface Player {
  id:                  number;
  fullName:            string;
  link:                string;
  firstName:           string;
  lastName:            string;
  primaryNumber:       string;
  birthDate:           Date;
  currentAge:          number;
  birthCity:           string;
  birthCountry:        Country;
  height:              string;
  weight:              number;
  active:              boolean;
  primaryPosition:     Position;
  useName:             string;
  middleName?:         string;
  boxscoreName:        string;
  nickName?:           string;
  gender:              AbstractGameCodeEnum;
  isPlayer:            boolean;
  isVerified:          boolean;
  draftYear?:          number;
  pronunciation?:      string;
  mlbDebutDate:        Date;
  batSide:             BatSide;
  pitchHand:           BatSide;
  nameFirstLast:       string;
  nameSlug:            string;
  firstLastName:       string;
  lastFirstName:       string;
  lastInitName:        string;
  initLastName:        string;
  fullFMLName:         string;
  fullLFMName:         string;
  strikeZoneTop:       number;
  strikeZoneBottom:    number;
  birthStateProvince?: string;
  nameMatrilineal?:    string;
  nameTitle?:          string;
}
// ...
export interface ProbablePitchers {
  away: OfficialScorer;
  home: OfficialScorer;
}

export interface Review {
  hasChallenges: boolean;
  away:          ReviewAway;
  home:          ReviewAway;
}

export interface ReviewAway {
  used:      number;
  remaining: number;
}

export interface Status {
  abstractGameState: string;
  codedGameState:    AbstractGameCodeEnum;
  detailedState:     string;
  statusCode:        AbstractGameCodeEnum;
  startTimeTBD:      boolean;
  abstractGameCode:  AbstractGameCodeEnum;
}

export interface GameDataTeams {
  away: PurpleAway;
  home: PurpleAway;
}
// ...
export interface Venue {
  id:        number;
  name:      string;
  link:      string;
  location:  Location;
  timeZone:  TimeZone;
  fieldInfo: FieldInfo;
  active:    boolean;
}
// ...
export interface Weather {
  condition: string;
  temp:      string;
  wind:      string;
}

export interface LiveData {
  plays:     Plays;
  linescore: Linescore;
  boxscore:  Boxscore;
  decisions: Decisions;
  leaders:   Leaders;
}

export interface Boxscore {
  teams:         BoxscoreTeams;
  officials:     Official[];
  info:          NoteElement[];
  pitchingNotes: any[];
}
// ...
export interface Decisions {
  winner: OfficialScorer;
  loser:  OfficialScorer;
  save:   OfficialScorer;
}

export interface Leaders {
  hitDistance: HitDistance;
  hitSpeed:    HitDistance;
  pitchSpeed:  HitDistance;
}

export interface Linescore {
  currentInning:        number;
  currentInningOrdinal: string;
  inningState:          string;
  inningHalf:           string;
  isTopInning:          boolean;
  scheduledInnings:     number;
  innings:              Inning[];
  teams:                LinescoreTeams;
  defense:              Defense;
  offense:              Offense;
  balls:                number;
  strikes:              number;
  outs:                 number;
}
// ...
export interface Plays {
  allPlays:      AllPlay[];
  currentPlay:   CurrentPlay;
  scoringPlays:  number[];
  playsByInning: PlaysByInning[];
}
// ...
export interface MetaData {
  wait:          number;
  timeStamp:     string;
  gameEvents:    string[];
  logicalEvents: string[];
}